+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [0.2] - Unreleased
 
+- `e` in keymap reads an SQL query from the minibuffer and displays the output in a temporary buffer.
+
 - Faces `pgmacs-table-header` and `pgmacs-table-data` are used to display the header and the rows of
   database tables.